Rename Language.Python.Syntax to Language.Python.DSL. Add a new Syntax module.
