+++ example/OptimizeTailRecursion.hs
@@ -19,7 +19,7 @@ import Data.Semigroup ((<>))
 
 import Language.Python.Optics
 import Language.Python.Internal.Syntax (CompoundStatement (..), Expr (..), Statement (..), SimpleStatement (..), SmallStatement (..), _Exprs, _Statements, argExpr, paramName)
-import Language.Python.Syntax
+import Language.Python.DSL
 import Language.Python.Syntax.Whitespace
 
 optimizeTailRecursion :: Raw Statement -> Maybe (Raw Statement)
